# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Launchpad API
  version: 1.0.0
servers:
  - url: http://localhost:4000/api/v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /auth/register:
    post:
      summary: Register user and organization
      requestBody:
        required: true
      responses:
        "201":
          description: Registered
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
      responses:
        "200":
          description: Authenticated
  /dashboard:
    get:
      summary: Dashboard metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Metrics
  /organizations/me:
    get:
      summary: Get current organization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Organization detail
    patch:
      summary: Update organization
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Updated
  /projects:
    get:
      summary: List projects
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Project list
    post:
      summary: Create project
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /projects/{projectId}:
    patch:
      summary: Update project
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete project
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /tasks:
    get:
      summary: List tasks
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Task list
    post:
      summary: Create task
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Created
  /tasks/{taskId}:
    patch:
      summary: Update task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated
    delete:
      summary: Delete task
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: taskId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted